requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==1.5.3
joblib==1.3.2
pyspark>=3.5.0
pymongo>=4.6.0
google-generativeai==0.3.1
